requirements.txt
# Claude_ML Home Run Prediction System
# Python Dependencies

# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0

# Machine Learning
scikit-learn>=1.1.0
xgboost>=1.6.0
imbalanced-learn>=0.9.0
joblib>=1.2.0

# Data Collection & APIs
requests>=2.28.0
beautifulsoup4>=4.11.0
pybaseball>=2.2.0
aiohttp>=3.8.0

# Scheduling & Async
schedule>=1.2.0

# Telegram Bot
python-telegram-bot>=20.0

# Weather API
openweather-python>=0.1.0

# Data Visualization (for backtesting charts)
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration & Logging
PyYAML>=6.0
python-dotenv>=0.19.0

# Date/Time handling
python-dateutil>=2.8.0
